Fix for issue #21. On MX Linux app only works once, then RF Explorer must be restarted.
This is just some kind of workaround. Normally the current configuration data is transmitted back from the device after it was set via the #0C2-F command. For unknown reasons on MX Linux the app only works once. When then restarted it states that the device cannot be found. The only way to make it work a gain is to restart the device (!) and the app. It looks like it helps to workaround the issue, by requesting config data explicitely, before setting it.